Fait: * 30pts) Ajouter des animations * animation de chute lorsque l'on place un jeton * animation de victoire qui qui fonctionne mais qui contient un bug
BUG:
* l'animation de victoire ne s'applique pas sur toute les cases qui donne la victoire.

* Calculage des points pour me retrouver facilement
effacer la partie courante automatiquement quand un joueur gagne (+10)
Activer/désactiver les en-têtes de la grille (+30)
Avertir l'usager qu'il faut se connecter (+30)
transition automatique vers la connexion, puis vers la partie (+10)
Ajouter des animations (+30)
total : 110/100

TODO
* trouver 20 points

+++ app/src/main/java/ca/cours5b5/stevendesroches/donnees/Disque.java
@@ -82,12 +82,15 @@ public final class Disque extends SourceDeDonnees {
 
     @Override
     public void detruireSauvegarde(String cheminSauvegarde) {
-
         File fichier = getFichier(cheminSauvegarde);
-        fichier.delete();
+
+        if(fichier.exists()) {
+            fichier.delete();
+        } else {
+            Log.d("detruireSauvegarde", this.getClass().getSimpleName()+"::Le fichier n");
+        }
 
     }
-
 
     private File getFichier(String cheminSauvegarde) {
 
@@ -98,7 +101,6 @@ public final class Disque extends SourceDeDonnees {
         return new File(repertoireRacine, nomFichier);
 
     }
-
 
     private String getNomFichier(String nomModele) {
